# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ACS Storage REST API
  version: "1"
  description: |
    Management and integration API for ACS Storage. The S3-compatible API is
    served separately (default port 9000) and follows the AWS S3 specification.

    **Authentication** — use a session cookie (web panel) or an access key:
    `Authorization: Bearer <ACCESS_KEY_ID>:<SECRET>` or HTTP Basic auth with the
    key ID as user name and the secret as password.

    **Errors** use RFC 9457 `application/problem+json`. POST and PATCH requests
    must send `Content-Type: application/json`.
servers:
  - url: /api/v1
security:
  - bearerKey: []
  - basicKey: []
  - session: []
components:
  securitySchemes:
    bearerKey: { type: http, scheme: bearer, description: "ACCESS_KEY_ID:SECRET" }
    basicKey: { type: http, scheme: basic }
    session: { type: apiKey, in: cookie, name: acs_session }
  parameters:
...
    id: { name: id, in: path, required: true, schema: { type: integer } }
...
  responses:
...
    NoContent: { description: No content }
  schemas:
...
    Share:
      type: object
      properties:
        id: { type: integer }
        token: { type: string }
        url: { type: string }
        type: { type: string, enum: [file, folder, upload] }
        bucket: { type: string }
        key: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        hasPassword: { type: boolean }
        maxDownloads: { type: integer }
        downloads: { type: integer }
        views: { type: integer }
        maxUploadBytes: { type: integer }
        note: { type: string }
        disabled: { type: boolean }
...
paths:
...
  /shares:
    get:
      summary: List share links (own, or all for admins)
      parameters: [{ name: bucket, in: query, schema: { type: string } }, { name: key, in: query, schema: { type: string } }]
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Share" } } } } } }
    post:
      summary: Create a share link
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type, bucket, key]
              properties:
                type: { type: string, enum: [file, folder, upload] }
                bucket: { type: string }
                key: { type: string, description: Object key, or prefix for folder/upload links }
                expiresAt: { type: string, format: date-time }
                password: { type: string }
                maxDownloads: { type: integer }
                maxUploadBytes: { type: integer }
                note: { type: string }
      responses: { "201": { description: Created } }
  /shares/{id}:
    parameters: [{ $ref: "#/components/parameters/id" }]
    patch:
      summary: Update a share link (send null to clear expiresAt, maxDownloads or password)
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { "200": { description: OK } }
    delete: { summary: Delete a share link, responses: { "204": { $ref: "#/components/responses/NoContent" } } }
  /public/shares/{token}:
    get:
      summary: Public share info (no auth)
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "410": { description: Expired, disabled or limit reached } }
...
  /public/shares/{token}/download:
    get:
      summary: Download from a file or folder share (download=1 counts toward the limit)
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string } }
        - { name: download, in: query, schema: { type: integer, enum: [1] } }
      responses: { "200": { description: Content } }
...